+++ lib/src/cursor.dart
@@ -3,6 +3,7 @@
 import 'package:flutter_android/android_database.dart' show Cursor, MatrixCursor;
 import 'package:meta/meta.dart' show experimental, required;
 
+import 'closable.dart' show SQLiteClosable;
 import 'exceptions.dart';
 
 /// Exposes results from a query on a [SQLiteDatabase].
@@ -12,7 +13,7 @@ import 'exceptions.dart';
 /// [SQLiteCursor].
 ///
 /// See: https://developer.android.com/reference/android/database/sqlite/SQLiteCursor
-class SQLiteCursor extends MatrixCursor {
+class SQLiteCursor extends MatrixCursor implements SQLiteClosable {
   /// Constructs an empty cursor.
   SQLiteCursor.empty()
     : super.empty();